reformat Pokemon cleaner, set seed back to int, impl Ditto special logic.

+++ src/main/java/cs6310/Pokemon/Bulbasaur.java
@@ -12,7 +12,7 @@ public class Bulbasaur extends Pokemon {
         setStats();
     }
 
-    public Bulbasaur(long seed) {
+    public Bulbasaur(int seed) {
         super(seed);
         setStats();
     }
